pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-incident-triage-env"
version = "0.1.0"
description = "Deterministic OpenEnv environment for SRE incident diagnosis and response."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "openai>=1.40.0",
  "openenv-core>=0.2.0",
  "pydantic>=2.8.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools]
packages = ["src"]
